Implement the use of '\' for line continuation in dependency files
Some of the lines in the dependency files, especially for the conditional
cases, can become quite long. This change implements 'line
continuation' for lines ending in '\'. The implementation is a bit
naive/straightforward. There is no real grammar behind the dependency
file parsing anyway, so this should not be too offensive.