Fail dependency parsing in case of undefined variables
In case an undefined variable is used in a dependency file, the
current implementation of generate-project stumbles on, after
showing a somewhat cryptic warning message.

This fix makes sure that parsing fails, and ipbb aborts with a clear
message about parsing errors, showing the exact lines, as well
as a message about which variable is undefined.